Flytter ut enumer for tiltakskode og arenatiltakskode for
