+++ lib/models/src/main/kotlin/no/nav/amt/lib/models/deltakerliste/tiltakstype/Tiltakskode.kt
@@ -1,0 +1,33 @@
+package no.nav.amt.lib.models.deltakerliste.tiltakstype
+
+enum class Tiltakskode {
+    ARBEIDSFORBEREDENDE_TRENING,
+    ARBEIDSRETTET_REHABILITERING,
+    AVKLARING,
+    DIGITALT_OPPFOLGINGSTILTAK,
+    GRUPPE_ARBEIDSMARKEDSOPPLAERING,
+    GRUPPE_FAG_OG_YRKESOPPLAERING,
+    JOBBKLUBB,
+    OPPFOLGING,
+    VARIG_TILRETTELAGT_ARBEID_SKJERMET,
+    HOYERE_UTDANNING,
+    AMO,
+    FAG_OG_YRKESOPPLAERING,
+    ;
+
+    fun toArenaKode() = when (this) {
+        ARBEIDSFORBEREDENDE_TRENING -> ArenaKode.ARBFORB
+        ARBEIDSRETTET_REHABILITERING -> ArenaKode.ARBRRHDAG
+        AVKLARING -> ArenaKode.AVKLARAG
+        DIGITALT_OPPFOLGINGSTILTAK -> ArenaKode.DIGIOPPARB
+        GRUPPE_ARBEIDSMARKEDSOPPLAERING -> ArenaKode.GRUPPEAMO
+        GRUPPE_FAG_OG_YRKESOPPLAERING -> ArenaKode.GRUFAGYRKE
+        JOBBKLUBB -> ArenaKode.JOBBK
+        OPPFOLGING -> ArenaKode.INDOPPFAG
+        VARIG_TILRETTELAGT_ARBEID_SKJERMET -> ArenaKode.VASV
+        AMO -> ArenaKode.ENKELAMO
+        FAG_OG_YRKESOPPLAERING -> ArenaKode.ENKFAGYRKE
+        HOYERE_UTDANNING -> ArenaKode.HOYEREUTD
+    }
+
+}
